requirements.txt

aiohttp>=3.12.15
alembic>=1.16.5
fastapi>=0.116.1
azure-storage-blob>=12.19.0
jinja2>=3.1.6
python-multipart>=0.0.20
python-slugify>=8.0.4
requests>=2.32.5
sqlmodel>=0.0.24
uvicorn>=0.35.0
